+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.0 (pending)
 
 * added `run.sh` to run locally
+* load books from /usr/local/yogi/etc/books
 
 
 ## 1.1.0 (2021-01-08)